% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countyFIPS.R
\name{countyFIPS}
\alias{countyFIPS}
\title{Identify counties for non-county locations}
\usage{
countyFIPS(lat, lon, location = NULL, place = NULL, key = NULL)
}
\arguments{
\item{lat}{The latitude coordinate for a location}

\item{lon}{The longitude coordinate for a location}

\item{location}{The name of a location given as a string}

\item{place}{The Place FIPS Code given as a string}

\item{key}{Your Google API key for geocoding location strings and place FIPS}
}
\value{
A dataframe returning the queried coordinates and resulting state, county, and county FIPS
}
\description{
\code{countyFIPS} converts lat/lon coordinates, location names, or
place FIPS codes to county-level FIPS codes.
}
\examples{
countyFIPS(lat = 39.9526, lon = -75.1652)
countyFIPS(location = "Philadelphia, PA", key = "XXXXXXXXXXXXXX")
countyFIPS(place = "4260000", key = "XXXXXXXXXXXXXX")
}
\keyword{API}
\keyword{FIPS,}
\keyword{counties,}
\keyword{geocoding,}
