% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_geocodes.R
\docType{data}
\name{place_geocodes}
\alias{place_geocodes}
\title{FIPS Codes and Location Names for Non Counties}
\format{A dataframe with 19518 rows and 2 columns:
 \describe{
   \item{fips}{FIPS code as character string}
   \item{full_name}{Name in "Location, State" format}
 }}
\source{
\url{https://www.census.gov/geographies/reference-files/2016/demo/popest/2016-fips.html}
}
\usage{
place_geocodes
}
\description{
A dataset constructed from 2016 Place and Consolidated City Codes provided by
US Census Bureau.
}
\keyword{datasets}
